+++ src/main/java/roomescape/domain/time/ReservationTimeRequestDTO.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.domain.time;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
